test: split version_cli and core_cli tests
